Add `return_index_map` functionality for functional interface
